services/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "twilio>=8.10.0",
    "vapi-python>=0.1.9",
    "vapi-server-sdk>=1.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.397",
    "ruff>=0.9.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]
